refactor: deviceSize 위치 변경
